requirements.txt.txt
python-telegram-bot==21.6
python-dotenv
openai
flask